refactor : root 프로젝트에 application 위치
